restructure code layout for dual backends
working tests as well
TODO: finish the new sketch conversion
