+++ src/renderer/metal/metal_scene.h
