+++ src/renderer/software/lineDraw.h
