+++ src/renderer/software/rasterizer.h
@@ -5,7 +5,7 @@
 #include <functional>
 
 #define MINITY_SCENE_TYPES_ONLY
-#include "scene.h"
+#include "../../freezer/old_scene.h"
 #include "simpleMath.h"
 #include "stats.h"
