+++ src/renderer/software/sdlHelpers.h
