+++ src/renderer/software/stats.h
