+++ src/renderer/software/triangleFill.h
